requirements.txt
autopep8==1.4.4
certifi==2019.6.16
chardet==3.0.4
Click==7.0
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Script==2.0.6
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pycodestyle==2.5.0
python-dotenv==0.10.3
requests==2.22.0
urllib3==1.25.3
visitor==0.1.3
Werkzeug==0.15.6
